refactored streaming types into separate module
